+++ lib/flutter_toast.dart
@@ -1,4 +1,4 @@
-library flutter_toast;
+library flutter_toast_notification;
 
 export 'src/flutter_toast.dart';
 export 'src/toast_position.dart';
@@ -6,4 +6,3 @@ export 'src/toast_type.dart';
 export 'src/toast_style.dart';
 export 'src/toast_style_type.dart';
 export 'src/icon_position.dart';
-