+++ test/fluttert_toast_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_toast/flutter_toast.dart';
+import 'package:flutter_toast_notification/flutter_toast.dart';
 
 void main() {
   group('FlutterToast', () {